Refactoring step 2 - introduction of common bevroad/utils.py
