+++ bevroad/mosaic.py
@@ -5,64 +5,13 @@ Stitches multiple BEV-warped images into a global map using pose metadata.
 """
 
 import argparse
-import json
-import math
 import os
 import sys
 
 import cv2
 import numpy as np
 
-
-def load_calibration(config_path):
-    if not os.path.exists(config_path):
-        print(f"Error: Calibration config file '{config_path}' does not exist.")
-        sys.exit(1)
-    try:
-        with open(config_path, 'r') as f:
-            return json.load(f)
-    except Exception as e:
-        print(f'Error loading calibration config {config_path}: {e}')
-        sys.exit(1)
-
-
-def parse_csv(csv_path):
-    if not os.path.exists(csv_path):
-        print(f"Error: Metadata CSV file '{csv_path}' does not exist.")
-        sys.exit(1)
-
-    records = []
-    with open(csv_path, 'r', encoding='utf-8') as f:
-        for line_num, line in enumerate(f, 1):
-            line = line.strip()
-            if not line:
-                continue
-            parts = line.split(',')
-            if len(parts) < 5:
-                print(f'Warning: line {line_num} in CSV is invalid and will be skipped.')
-                continue
-            try:
-                filename = parts[0]
-                ts = float(parts[1])
-                x = float(parts[2])
-                y = float(parts[3])
-                heading = float(parts[4])
-                records.append({'filename': filename, 'ts': ts, 'x': x, 'y': y, 'heading': heading})
-            except ValueError as e:
-                print(f'Warning: line {line_num} parsing failed ({e}). Skipped.')
-    return records
-
-
-def transform_point(x_local, y_local, robot_x, robot_y, heading):
-    """
-    Transforms local robot-frame coordinates (X=forward, Y=left)
-    to global coordinates based on robot's position and heading (radians).
-    """
-    c = math.cos(heading)
-    s = math.sin(heading)
-    x_global = robot_x + x_local * c - y_local * s
-    y_global = robot_y + x_local * s + y_local * c
-    return x_global, y_global
+from bevroad.utils import load_calibration, parse_csv, transform_point
 
 
 def main():
